Refactor PledgeSlider component and remove PledgeSliderV2
- Replaced the PledgeSliderV2 component with an updated PledgeSlider that utilizes the Embla carousel for improved performance and responsiveness.
- Enhanced the PledgeSlider to handle image rendering and selection more efficiently, including responsive updates with a resize observer.
- Cleaned up the codebase by deleting the obsolete PledgeSliderV2 file.

This commit focuses on optimizing the pledge selection experience and streamlining the component structure.
